requirements.txt
numpy==1.23.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.1.2
widgetsnbextension==4.0.14
wrapt==1.17.2
Flask==2.1.2
scikit-learn==1.6.1
tensorflow==2.16.2
keras==3.8.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
python-dotenv
python-decouple
protobuf==3.20.3
firebase-admin==6.6.0
gunicorn==20.1.0  # Asegúrate de incluir Gunicorn
google-cloud-storage
joblib
requests==2.32.3
flask-cors==5.0.1
fastapi
uvicorn
python-multipart
jinja2
email-validator
